sentryprime-backend-fixed/requirements.txt
beautifulsoup4==4.13.4
blinker==1.9.0
cachelib==0.13.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
Flask==3.1.1
Flask-Caching==2.3.1
flask-cors==6.0.0
Flask-SQLAlchemy==3.1.1
greenlet==3.2.4
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
lxml==6.0.0
MarkupSafe==3.0.2
openai==1.54.4
psycopg2-binary==2.9.9
pydantic==2.10.3
PyJWT==2.8.0
requests==2.32.4
soupsieve==2.7
SQLAlchemy==2.0.41
stripe==12.4.0
typing_extensions==4.14.0
urllib3==2.5.0
Werkzeug==3.1.3
